io_antarctica_scene: Track exporter works on basic tracks without fancy features
Tracks with cannons or goals are failing to fully export for some reason.
Additional changes include:
* io_scene_spm: Fix events where it is possible for Nonetype object to be read
* Most classes moved from track exporter into a special utility file
* Track exporter reads the proper location for the assets (data) path
* Track exporter makes full use of built-in Blender logging facilities
* Kart and track exporters have reworded buttons for better consistency
* Clean up some unused imports
