+++ io_antarctica_scene/stk_utils.py
@@ -159,6 +159,32 @@ def selectObjectsInList(obj_list):
         if not obj.select_get():
             obj.select_set(True)
 
+def searchNodeTreeForImage(node_tree, uv_num):
+    # Check if there is a node tree
+    # If so, search the STK shader node for an image
+    if node_tree is not None:
+        try:
+            shader_node = node_tree.nodes['Principled BSDF']
+            if shader_node.inputs['Base Color'].is_linked:
+                # Get the connected node
+                child = shader_node.inputs['Base Color'].links[0].from_node
+                if type(child) is bpy.types.ShaderNodeTexImage and uv_num == 1:
+                    return os.path.basename(child.image.filepath)
+                elif type(child) is bpy.types.ShaderNodeMixRGB:
+                    uvOne = child.links['Color1'].from_node
+                    uvTwo = child.links['Color2'].from_node
+                    if type(uvOne) is bpy.types.ShaderNodeTexImage and uv_num == 1:
+                        return os.path.basename(uvOne.image.filepath)
+                    if type(uvTwo) is bpy.types.ShaderNodeTexImage and uv_num == 2:
+                        return os.path.basename(uvTwo.image.filepath)
+            else:
+                #print("Texture node not found, skipping this input node")
+                return ""
+        except:
+            return ""
+    else:
+        return ""
+
 # ------------------------------------------------------------------------------
 #! Utility function, creates all properties in a given object
 #!
